Create collecting_apples.cpp
Consider the “Collecting Apples” problem given in class, rewrite the solution such that moving from start to end includes three types of moves:

Right, Down, Right-Down.  The Right-down move means that you can move diagonally.

Also, black apples indicate that these are blocked and cannot be crossed as part of the path from start to end. The blocked cells are represented by negative values in the input.


Write the function CountingApples that takes as input a 2D array apples, and the values of two integers 
�
N and 
�
M (
2
≤
�
,
 
2
≤
�
)
2≤N, 2≤M). 
�
N represents the number of rows, and 
�
M represents the number of columns in the given apples 2D array.

You will not be asked to read the values, the apples 2D array will be ready automatically and fed to your CountingApples function. 

If there is no path from start to end, the maximum number of collected apples will be 0. Also, if the start cell is a blocked one, there is no solution and the maximum number of collected apples will be 0 in this case as well.