Create minimize.cpp
A company ABC requires to use an online server for research experimentation for a certain number of days. The server will be used on certain days but not every day.

To use an online server, the company is required to subscribe to a cloud provider. A cloud provider offers 3 subscriptions options: 1-day subscription, a 1-week subscription, or a 30-day subscription, each one of them has a price that can be found in a given array of prices, where prices[0] represents the price of a 1-day subscription, prices[1] represents the price of a 1-week subscription, and prices[2] represents the price of a 30-days subscription.

Write a function MinSubscription to find the minimum amount of money that company ABC needs to have in order to use a server during all the experimentation days.

Your function MinSubscription takes as parameters 3 values:

E  which represents the number of experimentation days.

Days which has  E integer values that represent the experimentation days Di where  1< Di <=365

Prices which has 3 values: 1-day Subscription, 1-Week Subscription, and 30-day Subscription