Create dfs.cpp
Consider the LinkedIn solution that allows people to connect on a social media platform for employment purposes. Every user on LinkedIn can establish one or more connections with other users on the same platform. In addition to the connections, assume that LinkedIn also tracks and saves the employer company of each user.

Write a function that allows a user to discover how connected he/she is to some company. The function targets a specific input company and prints how far the user is in his connections from that company, and how many users he is connected to from that company at that level.